// GoogleMapsGrid/generated/ManifestTypes.d.ts
/*
*This is auto generated from the ControlManifest.Input.xml file
*/

// Define IInputs and IOutputs Type. They should match with ControlManifest.
export interface IInputs {
    GoogleMapsApiKey: ComponentFramework.PropertyTypes.StringProperty;
    LatitudeField: ComponentFramework.PropertyTypes.StringProperty;
    LongitudeField: ComponentFramework.PropertyTypes.StringProperty;
    TitleField: ComponentFramework.PropertyTypes.StringProperty;
    PrimaryMarkerShape: ComponentFramework.PropertyTypes.StringProperty;
    MarkerStyleMapping: ComponentFramework.PropertyTypes.StringProperty;
    ColorOptionSetField: ComponentFramework.PropertyTypes.StringProperty;
    MapHeight: ComponentFramework.PropertyTypes.WholeNumberProperty;
    dataSet: ComponentFramework.PropertyTypes.DataSet;
}
export interface IOutputs {
}